Move promisifyEvent to main utils file
