+++ tests/jest/BookReader/utils.test.js
@@ -1,5 +1,5 @@
 import sinon from 'sinon';
-import { afterEventLoop } from '../utils.js';
+import { afterEventLoop, eventTargetMixin } from '../utils.js';
 import {
   clamp,
   cssPercentage,
@@ -12,6 +12,7 @@ import {
   poll,
   polyfillCustomEvent,
   PolyfilledCustomEvent,
+  promisifyEvent,
   sleep,
 } from '@/src/BookReader/utils.js';
 
@@ -182,5 +183,35 @@ describe('sleep', () => {
 
     await afterEventLoop();
     expect(spy.callCount).toBe(1);
+  });
+});
+
+describe('promisifyEvent', () => {
+  test('Resolves once event fires', async () => {
+    const fakeTarget = eventTargetMixin();
+    const resolveSpy = sinon.spy();
+    promisifyEvent(fakeTarget, 'pause').then(resolveSpy);
+
+    await afterEventLoop();
+    expect(resolveSpy.callCount).toBe(0);
+    fakeTarget.dispatchEvent('pause', {});
+    await afterEventLoop();
+    expect(resolveSpy.callCount).toBe(1);
+  });
+
+  test('Only resolves once', async () => {
+    const fakeTarget = eventTargetMixin();
+    const resolveSpy = sinon.spy();
+    promisifyEvent(fakeTarget, 'pause').then(resolveSpy);
+
+    await afterEventLoop();
+    expect(resolveSpy.callCount).toBe(0);
+    fakeTarget.dispatchEvent('pause', {});
+    fakeTarget.dispatchEvent('pause', {});
+    fakeTarget.dispatchEvent('pause', {});
+    fakeTarget.dispatchEvent('pause', {});
+
+    await afterEventLoop();
+    expect(resolveSpy.callCount).toBe(1);
   });
 });
